vision/requirements.txt
Pillow
ultralytics
numpy
opencv-python